Monsters can now convey any attribute score. Ogres and trolls convey CON, giants STR (as before), snakes DEX, mind flayers and gray dragons INT, angels, owlbears, and silver dragons WIS, and nympsh CHA. The attribute exercise system has been deactivated.

+++ include/permonst.h
@@ -59,8 +59,8 @@ struct permonst {
 		intl,					/* intelligence */
 		wis,					/* wisdom */
 		cha;                    /* charisma */
-	unsigned short mresists;             /* resistances */
-	unsigned short mconveys;             /* conveyed by eating */
+	unsigned long mresists;             /* resistances */
+	unsigned long mconveys;             /* conveyed by eating */
     unsigned long mflags1,      /* boolean bitflags */
         mflags2, mflags3;       /* more boolean bitflags */
     uchar difficulty;		/* toughness (formerly from  makedefs -m) */
